refactor: refactor tx list of txs page and block page
use graphql to fetch data and unify tx list component in
txs/block/account page
